Fix errors with paths containing spaces
When invoking bundletool with a path containing spaces, it failed
because the command did not take into account that the path may contain
spaces.

When passing a path with a space on it, the shell interpreted the
space as a separation between parameters:
`bundletool --bundle /path/with/spaces on/it`

As the substring 'on/it' is considered a second parameter but because
it does not starts with '--' it generates the error:
"Error while parsing the flags: Syntax error: flags should start with --"

The solution is to add double quotes around the path variables when generating
the command string used to invoke bundletool or any other shell command.